feat(wkb): extract dependency-free WKB and WKT codecs
