+++ modules/geoarrow/tsconfig.json
@@ -7,6 +7,7 @@
     {"path": "../types"},
     {"path": "../crs"},
     {"path": "../core"},
-    {"path": "../polygon"}
+    {"path": "../polygon"},
+    {"path": "../wkb"}
   ]
 }
