+++ modules/wkb/package.json
@@ -1,6 +1,6 @@
 {
-  "name": "@math.gl/geoarrow",
-  "description": "Arrow-independent GeoArrow physical layouts and geometry kernels",
+  "name": "@math.gl/wkb",
+  "description": "Dependency-free WKB and WKT geometry codecs",
   "license": "MIT",
   "type": "module",
   "publishConfig": {
@@ -8,10 +8,10 @@
   },
   "version": "5.0.0-alpha.1",
   "keywords": [
-    "geoarrow",
+    "wkb",
+    "wkt",
+    "geometry",
     "geospatial",
-    "columnar",
-    "typed-array",
     "math"
   ],
   "repository": {
@@ -26,22 +26,11 @@
       "types": "./dist/index.d.ts",
       "import": "./dist/index.js",
       "require": "./dist/index.cjs"
-    },
-    "./worker": {
-      "types": "./dist/worker.d.ts",
-      "import": "./dist/worker.js",
-      "require": "./dist/worker.cjs"
     }
   },
   "files": [
     "dist",
     "src"
   ],
-  "sideEffects": false,
-  "dependencies": {
-    "@math.gl/crs": "5.0.0-alpha.1",
-    "@math.gl/polygon": "5.0.0-alpha.1",
-    "@math.gl/types": "5.0.0-alpha.1",
-    "@math.gl/wkb": "5.0.0-alpha.1"
-  }
+  "sideEffects": false
 }
