+++ modules/wkb/src/index.ts
@@ -1,0 +1,14 @@
+// math.gl
+// SPDX-License-Identifier: MIT
+// Copyright (c) vis.gl contributors
+
+export type {WellKnownDimension, WellKnownGeometry} from './types';
+export {
+  getWellKnownDimensionSize,
+  inferWellKnownGeometryDimension
+} from './types';
+
+export type {WKBParseOptions, WKBParseResult} from './wkb';
+export {parseWKB, writeWKB} from './wkb';
+
+export {formatWKT, parseWKT} from './wkt';
