+++ modules/wkb/tsconfig.json
@@ -1,0 +1,6 @@
+{
+  "extends": "../../tsconfig.json",
+  "include": ["src/**/*"],
+  "exclude": ["node_modules"],
+  "compilerOptions": {"composite": true, "rootDir": "src", "outDir": "dist"}
+}
